Extract headers into separate module
